refactor: convert capabilities:create command to yargs
