+++ src/commands/index.ts
@@ -12,6 +12,7 @@ import appsSettingsCommand from './apps/settings.js'
 import appsSettingsUpdateCommand from './apps/settings/update.js'
 import appsUpdateCommand from './apps/update.js'
 import capabilitiesCommand from './capabilities.js'
+import capabilitiesCreateCommand from './capabilities/create.js'
 import capabilitiesNamespacesCommand from './capabilities/namespaces.js'
 import capabilitiesPresentationCommand from './capabilities/presentation.js'
 import capabilitiesPresentationCreateCommand from './capabilities/presentation/create.js'
@@ -55,6 +56,7 @@ export const commands: CommandModule<object, any>[] = [
 	appsSettingsUpdateCommand,
 	appsUpdateCommand,
 	capabilitiesCommand,
+	capabilitiesCreateCommand,
 	capabilitiesNamespacesCommand,
 	capabilitiesPresentationCommand,
 	capabilitiesPresentationCreateCommand,
